Maintenance work
* String arrays ProcessErrorStrings, ProcessStateStrings, and ExitStatusStrings
have been moved from shepherd.cpp and slicer.cpp into strings.cpp, which
contains the implementation of three ToString methods, one for each of types
QProcess::ProcessError, QProcess::ProcessState, and QProcess::ExitStatus.

* Clean up some debugging messages.

* Classes Shepherd and Slicer: In order for the code constructing an instance
to have an opportunity to connect signals to slots before the process is
actually launched, these classes' constructors now do everything *but* invoke
QProcess::start(), which has now been moved into a new method called, appropriately
enough, start. This new method also makes sure that the process is not running
before it attempts to start it.

* Adjust code that constructs an object, configures it, and connects slots to
its signals so that the signal-connecting is done last, after configuration.

* Window.cpp: Add new enum class TabIndex, to replace hard-coded constants, the
tab indices, with symbolic names. (The operator+ overload for TabIndex is a trick
to get around a quirk of the C++ type system.)