Adjust when the print speed is configured according to the job.
Previously, the very first thing we did when we started printing a job was to change the print speed to the speed defined in the job object. However, this means the raise before print solution dispensing, the lower after print solution dispensing, and the raise after printing is complete are *also* done at the job's print speed, introducing unnecessary tedium and delay.

Consequently, the A-series of steps ("before printing") was rearranged, and a step split in two since the lowering of the build platform is now done in two steps -- fast from 70 mm to 50 mm, then job speed from 50 mm to first layer height. It was also necessary to split step C1 (previously the only "after printing" step) in two, and *that* necessitated changes to how aborts are handled.