Mountmon: MediaAvailable==FALSE issue
Sometimes, when a USB stick is inserted, the properties for the Drive object we receive from udisks2 has the MediaAvailable property set to FALSE. I've added code to dump all the object's properties into the debugging log. Also, the code now warns if MediaAvailable==FALSE, but will try mounting it anyway.